requirements.txt
discord.py==1.7.3
Flask==2.0.2
gunicorn==20.1.0

